Fixed floating point accuracy issue with spot lights on some platforms.
https://fogbugz.unity3d.com/f/cases/1416456/
https://unity3d.atlassian.net/browse/TRND-2072

This fixes incorrect distance falloff for spot and point lights on Nintendo Switch.

+++ Packages/com.unity.render-pipelines.universal/ShaderLibrary/RealtimeLights.hlsl
@@ -13,7 +13,7 @@ struct Light
 {
     half3   direction;
     half3   color;
-    half    distanceAttenuation;
+    float   distanceAttenuation; // full-float precision required on some platforms
     half    shadowAttenuation;
     uint    layerMask;
 };
@@ -161,7 +161,8 @@ Light GetAdditionalPerObjectLight(int perObjectLightIndex, float3 positionWS)
     float distanceSqr = max(dot(lightVector, lightVector), HALF_MIN);
 
     half3 lightDirection = half3(lightVector * rsqrt(distanceSqr));
-    half attenuation = half(DistanceAttenuation(distanceSqr, distanceAndSpotAttenuation.xy) * AngleAttenuation(spotDirection.xyz, lightDirection, distanceAndSpotAttenuation.zw));
+    // full-float precision required on some platforms
+    float attenuation = DistanceAttenuation(distanceSqr, distanceAndSpotAttenuation.xy) * AngleAttenuation(spotDirection.xyz, lightDirection, distanceAndSpotAttenuation.zw);
 
     Light light;
     light.direction = lightDirection;
